+++ frontend/src/lib/components/ChildrenRegistration.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
 	import { base } from '$app/paths';
-	import Breadcrumbs from '$lib/components//Breadcrumbs.svelte';
+	import Breadcrumbs from '$lib/components/Navigation/Breadcrumbs.svelte';
 	import AlertMessage from '$lib/components/AlertMessage.svelte';
 	import DataInput from '$lib/components/DataInput/DataInput.svelte';
 	import NavigationButtons from '$lib/components/Navigation/NavigationButtons.svelte';
